requirements.txt
psd-tools
Pillow
PyYAML
pytest
